+++ sdk/messaging/azeventhubs/event_data_batch_unit_test.go
@@ -108,7 +108,7 @@ func TestUnitEventDataBatchUnitTests(t *testing.T) {
 		}, nil)
 
 		require.NoError(t, err)
-		require.EqualValues(t, 1, mb.NumMessages())
+		require.EqualValues(t, 1, mb.NumEvents())
 		require.EqualValues(t, 172, mb.NumBytes())
 
 		actualBytes, err := mb.toAMQPMessage().MarshalBinary()
@@ -132,7 +132,7 @@ func TestUnitEventDataBatchUnitTests(t *testing.T) {
 		}, nil)
 
 		require.NoError(t, err)
-		require.EqualValues(t, 1, mb.NumMessages())
+		require.EqualValues(t, 1, mb.NumEvents())
 		require.EqualValues(t, 4357, mb.NumBytes())
 
 		actualBytes, err := mb.toAMQPMessage().MarshalBinary()
@@ -169,7 +169,7 @@ func TestUnitEventDataBatchUnitTests(t *testing.T) {
 		require.EqualValues(t, 121, mb.currentSize)
 
 		sizeBefore := mb.NumBytes()
-		countBefore := mb.NumMessages()
+		countBefore := mb.NumEvents()
 
 		err = mb.AddEventData(&EventData{
 			Body: as2k[:],
@@ -177,7 +177,7 @@ func TestUnitEventDataBatchUnitTests(t *testing.T) {
 		require.EqualError(t, err, ErrEventDataTooLarge.Error())
 
 		require.Equal(t, sizeBefore, mb.NumBytes(), "size is unchanged when a message fails to get added")
-		require.Equal(t, countBefore, mb.NumMessages(), "count is unchanged when a message fails to get added")
+		require.Equal(t, countBefore, mb.NumEvents(), "count is unchanged when a message fails to get added")
 	})
 
 	t.Run("addConcurrently", func(t *testing.T) {
@@ -200,7 +200,7 @@ func TestUnitEventDataBatchUnitTests(t *testing.T) {
 		}
 
 		wg.Wait()
-		require.EqualValues(t, 100, mb.NumMessages())
+		require.EqualValues(t, 100, mb.NumEvents())
 	})
 }